phase-5/services/notification/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlmodel>=0.0.22
asyncpg>=0.30.0
python-dotenv>=1.0.0
